+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## 0.7
 
 - Added support for django 1.10 and 1.11
+- Dropped support for django 1.3 - 1.7
 - Fixed issue where django version would be forced to 1.8 when installing django-parsley. More details <a href="https://github.com/agiliq/Django-parsley/issues/72" target="_blank">here</a>.
+- Added parsleyfy template tag
 
 ## 0.6